ospf6d: Move non-NSSA ABR code from ospf6_nssa.c to ospf6_abr.c
Move general ABR infrastructure code not specific to NSSA support
(ospf6_abr_task() and friends) out of ospf6_nssa.c. It proably got copied in
there (from ospfd) when adding NSSA support since it was a prerequisite for
that. Now that general intera-area summary orgination uses the ABR task as
well, it no longer makes sense to hide it in ospf6_nssa.c, so move it to
ospf6_abr.c.

While we're at it, also clean up some other unused cruft from
ospf6_nssa.{c,h} and move things which should not be private (like LSA
flags) to their proper place.
