+++ ospf6d/ospf6_flood.c
@@ -23,6 +23,7 @@
 #include "ospf6_area.h"
 #include "ospf6_interface.h"
 #include "ospf6_neighbor.h"
+#include "ospf6_abr.h"
 
 #include "ospf6_flood.h"
 #include "ospf6_nssa.h"
