+++ ospf6d/ospf6_intra.c
@@ -29,7 +29,6 @@
 #include "ospf6_intra.h"
 #include "ospf6_asbr.h"
 #include "ospf6_abr.h"
-#include "ospf6_nssa.h"
 #include "ospf6_flood.h"
 #include "ospf6d.h"
 #include "ospf6_spf.h"
