+++ ospf6d/ospf6_lsa.h
@@ -248,6 +248,7 @@ struct ospf6_lsa {
 #define OSPF6_LSA_UNAPPROVED 0x10
 #define OSPF6_LSA_SEQWRAPPED 0x20
 #define OSPF6_LSA_FLUSH      0x40
+#define OSPF6_LSA_LOCAL_XLT  0x80
 
 struct ospf6_lsa_handler {
 	uint16_t lh_type; /* host byte order */
